deduplicate webhook code with class inheritance
